requirements.txt
fastmcp
python-fortimanagerapi
python-dotenv 